Split binary numeric execution into a module tree
Mechanical move mirroring the compare split: dtype-generic lane machinery
stays in numeric/mod.rs (bounds relaxed to T: Default so wider decimal types
can reuse it), primitive-specific code moves to numeric/primitive.rs, tests
to numeric/tests.rs.
