+++ vortex-array/src/scalar_fn/fns/binary/numeric/primitive.rs
@@ -10,6 +10,11 @@ use vortex_error::vortex_err;
 use vortex_mask::AllOr;
 use vortex_mask::Mask;
 
+use super::CheckedValues;
+use super::any_valid_error;
+use super::check_numeric_errors;
+use super::checked_all_lanes;
+use super::checked_valid_lanes;
 use crate::ArrayRef;
 use crate::ExecutionCtx;
 use crate::IntoArray;
@@ -104,8 +109,8 @@ impl<T: CheckedArithmetic> CheckedPrimitiveOp<T> for CheckedDiv {
     }
 }
 
-/// Execute a numeric operation between two arrays.
-pub(crate) fn execute_numeric(
+/// Execute a numeric operation between two primitive-typed arrays.
+pub(super) fn execute_numeric_primitive(
     lhs: &ArrayRef,
     rhs: &ArrayRef,
     op: NumericOperator,
@@ -214,7 +219,7 @@ where
 
 /// A primitive binary-operator operand: a materialized buffer, a non-null constant, or an
 /// all-null constant.
-pub(super) enum PrimitiveOperand<T: NativePType> {
+pub(crate) enum PrimitiveOperand<T: NativePType> {
     Array {
         values: Buffer<T>,
         validity: Validity,
@@ -228,7 +233,7 @@ pub(super) enum PrimitiveOperand<T: NativePType> {
 }
 
 impl<T: NativePType> PrimitiveOperand<T> {
-    pub(super) fn try_new(array: &ArrayRef, ctx: &mut ExecutionCtx) -> VortexResult<Self> {
+    pub(crate) fn try_new(array: &ArrayRef, ctx: &mut ExecutionCtx) -> VortexResult<Self> {
         if let Some(constant) = array.as_opt::<Constant>() {
             return Ok(
                 match constant.scalar().as_primitive().try_typed_value::<T>()? {
@@ -252,39 +257,18 @@ impl<T: NativePType> PrimitiveOperand<T> {
         Ok(Self::Array { values, validity })
     }
 
-    pub(super) fn len(&self) -> usize {
+    pub(crate) fn len(&self) -> usize {
         match self {
             Self::Array { values, .. } => values.len(),
             Self::Constant { len, .. } | Self::Null(len) => *len,
         }
     }
 
-    pub(super) fn validity(&self) -> Validity {
+    pub(crate) fn validity(&self) -> Validity {
         match self {
             Self::Array { validity, .. } => validity.clone(),
             Self::Constant { validity, .. } => validity.clone(),
             Self::Null(_) => Validity::AllInvalid,
-        }
-    }
-}
-
-struct CheckedValues<T: NativePType> {
-    values: Buffer<T>,
-    failed: bool,
-}
-
-impl<T: NativePType> CheckedValues<T> {
-    fn zeroed(len: usize) -> Self {
-        Self {
-            values: Buffer::<T>::zeroed(len),
-            failed: false,
-        }
-    }
-
-    fn failed(len: usize) -> Self {
-        Self {
-            values: Buffer::<T>::zeroed(len),
-            failed: true,
         }
     }
 }
@@ -495,133 +479,6 @@ where
     Op: CheckedPrimitiveOp<T>,
 {
     checked_valid_lanes(rhs.len(), valid_bits, |idx| Op::checked(lhs, rhs[idx]))
-}
-
-// Checked one-pass ops delay early exit until the end of a small block. This
-// keeps the loop generic while avoiding a branch-driven exit decision on every
-// lane; it is deliberately independent of mask density or input length.
-const CHECKED_BLOCK_LANES: usize = 16;
-
-fn checked_all_lanes<T, F>(len: usize, mut checked_at: F) -> CheckedValues<T>
-where
-    T: NativePType,
-    F: FnMut(usize) -> Option<T>,
-{
-    let mut values = BufferMut::<T>::with_capacity(len);
-    let mut base = 0;
-
-    while base + CHECKED_BLOCK_LANES <= len {
-        let mut block_failed = false;
-        for idx in base..base + CHECKED_BLOCK_LANES {
-            match checked_at(idx) {
-                Some(value) => {
-                    // SAFETY: the buffer is allocated with capacity `len`, and
-                    // this loop pushes at most one value for each `idx`.
-                    unsafe { values.push_unchecked(value) };
-                }
-                None => {
-                    block_failed = true;
-                    // SAFETY: the buffer is allocated with capacity `len`, and
-                    // this loop pushes at most one value for each `idx`.
-                    unsafe { values.push_unchecked(T::default()) };
-                }
-            }
-        }
-
-        if block_failed {
-            return CheckedValues::failed(len);
-        }
-        base += CHECKED_BLOCK_LANES;
-    }
-
-    for idx in base..len {
-        let Some(value) = checked_at(idx) else {
-            return CheckedValues::failed(len);
-        };
-        // SAFETY: the buffer is allocated with capacity `len`, and this loop
-        // pushes at most one value for each `idx`.
-        unsafe { values.push_unchecked(value) };
-    }
-
-    CheckedValues {
-        values: values.freeze(),
-        failed: false,
-    }
-}
-
-fn checked_valid_lanes<T, F>(
-    len: usize,
-    valid_bits: &BitBuffer,
-    mut checked_at: F,
-) -> CheckedValues<T>
-where
-    T: NativePType,
-    F: FnMut(usize) -> Option<T>,
-{
-    let mut values = BufferMut::<T>::zeroed(len);
-    let mut failed = false;
-    {
-        let values = values.as_mut_slice();
-        for_each_valid_idx(len, valid_bits, |idx| {
-            let Some(value) = checked_at(idx) else {
-                failed = true;
-                return false;
-            };
-            values[idx] = value;
-            true
-        });
-    }
-
-    CheckedValues {
-        values: values.freeze(),
-        failed,
-    }
-}
-
-fn any_valid_error<F>(len: usize, valid_bits: &BitBuffer, is_error: F) -> bool
-where
-    F: Fn(usize) -> bool,
-{
-    !for_each_valid_idx(len, valid_bits, |idx| !is_error(idx))
-}
-
-fn for_each_valid_idx<F>(len: usize, valid_bits: &BitBuffer, mut f: F) -> bool
-where
-    F: FnMut(usize) -> bool,
-{
-    debug_assert_eq!(len, valid_bits.len());
-
-    for (word_idx, valid_word) in valid_bits.chunks().iter_padded().enumerate() {
-        if valid_word == 0 {
-            continue;
-        }
-
-        let offset = word_idx * 64;
-        let lanes = len.saturating_sub(offset).min(64);
-        if lanes == 64 && valid_word == u64::MAX {
-            for bit_idx in 0..64 {
-                if !f(offset + bit_idx) {
-                    return false;
-                }
-            }
-            continue;
-        }
-
-        let mut valid_word = if lanes == 64 {
-            valid_word
-        } else {
-            valid_word & ((1u64 << lanes) - 1)
-        };
-        while valid_word != 0 {
-            let bit_idx = valid_word.trailing_zeros() as usize;
-            if !f(offset + bit_idx) {
-                return false;
-            }
-            valid_word &= valid_word - 1;
-        }
-    }
-
-    true
 }
 
 trait CheckedArithmetic: NativePType {
@@ -913,224 +770,3 @@ impl_checked_signed!(i16, widening_mul: i32);
 impl_checked_signed!(i32, widening_mul: i64);
 impl_checked_signed!(i64, overflowing_mul);
 impl_checked_float!(f16, f32, f64);
-
-fn check_numeric_errors(failed: bool, error: &'static str) -> VortexResult<()> {
-    if failed {
-        return Err(vortex_err!(InvalidArgument: "{}", error));
-    }
-
-    Ok(())
-}
-#[cfg(test)]
-mod test {
-    use vortex_buffer::buffer;
-    use vortex_error::VortexResult;
-
-    use crate::ArrayRef;
-    use crate::IntoArray;
-    use crate::RecursiveCanonical;
-    use crate::VortexSessionExecute;
-    use crate::array_session;
-    use crate::arrays::ConstantArray;
-    use crate::arrays::PrimitiveArray;
-    use crate::assert_arrays_eq;
-    use crate::builtins::ArrayBuiltins;
-    use crate::scalar::Scalar;
-    use crate::scalar_fn::fns::operators::Operator;
-    use crate::validity::Validity;
-
-    fn sub_scalar(array: &ArrayRef, scalar: impl Into<Scalar>) -> VortexResult<ArrayRef> {
-        array
-            .binary(
-                ConstantArray::new(scalar, array.len()).into_array(),
-                Operator::Sub,
-            )
-            .and_then(|a| {
-                a.execute::<RecursiveCanonical>(&mut array_session().create_execution_ctx())
-            })
-            .map(|a| a.0.into_array())
-    }
-
-    #[test]
-    fn test_scalar_subtract_unsigned() {
-        let mut ctx = array_session().create_execution_ctx();
-        let values = buffer![1u16, 2, 3].into_array();
-        let result = sub_scalar(&values, 1u16).unwrap();
-        assert_arrays_eq!(result, PrimitiveArray::from_iter([0u16, 1, 2]), &mut ctx);
-    }
-
-    #[test]
-    fn test_scalar_subtract_signed() {
-        let mut ctx = array_session().create_execution_ctx();
-        let values = buffer![1i64, 2, 3].into_array();
-        let result = sub_scalar(&values, -1i64).unwrap();
-        assert_arrays_eq!(result, PrimitiveArray::from_iter([2i64, 3, 4]), &mut ctx);
-    }
-
-    #[test]
-    fn test_scalar_subtract_nullable() {
-        let mut ctx = array_session().create_execution_ctx();
-        let values = PrimitiveArray::from_option_iter([Some(1u16), Some(2), None, Some(3)]);
-        let result = sub_scalar(&values.into_array(), Some(1u16)).unwrap();
-        assert_arrays_eq!(
-            result,
-            PrimitiveArray::from_option_iter([Some(0u16), Some(1), None, Some(2)]),
-            &mut ctx
-        );
-    }
-
-    #[test]
-    fn test_scalar_subtract_float() {
-        let mut ctx = array_session().create_execution_ctx();
-        let values = buffer![1.0f64, 2.0, 3.0].into_array();
-        let result = sub_scalar(&values, -1f64).unwrap();
-        assert_arrays_eq!(
-            result,
-            PrimitiveArray::from_iter([2.0f64, 3.0, 4.0]),
-            &mut ctx
-        );
-    }
-
-    #[test]
-    fn test_scalar_subtract_float_underflow_is_ok() {
-        let values = buffer![f32::MIN, 2.0, 3.0].into_array();
-        let _results = sub_scalar(&values, 1.0f32).unwrap();
-        let _results = sub_scalar(&values, f32::MAX).unwrap();
-    }
-
-    #[test]
-    fn test_float_divide_by_zero_is_ok() {
-        let mut ctx = array_session().create_execution_ctx();
-        let values = buffer![1.0f64, -1.0].into_array();
-        let result = values
-            .binary(
-                ConstantArray::new(0.0f64, values.len()).into_array(),
-                Operator::Div,
-            )
-            .and_then(|a| a.execute::<PrimitiveArray>(&mut array_session().create_execution_ctx()))
-            .unwrap();
-
-        assert_arrays_eq!(
-            result,
-            PrimitiveArray::from_iter([f64::INFINITY, f64::NEG_INFINITY]),
-            &mut ctx
-        );
-    }
-
-    #[test]
-    fn test_integer_overflow_errors() {
-        let values = buffer![u8::MAX].into_array();
-        let result = values
-            .binary(
-                ConstantArray::new(1u8, values.len()).into_array(),
-                Operator::Add,
-            )
-            .and_then(|a| a.execute::<PrimitiveArray>(&mut array_session().create_execution_ctx()));
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_integer_divide_by_zero_errors() {
-        let values = buffer![1i32].into_array();
-        let result = values
-            .binary(
-                ConstantArray::new(0i32, values.len()).into_array(),
-                Operator::Div,
-            )
-            .and_then(|a| a.execute::<PrimitiveArray>(&mut array_session().create_execution_ctx()));
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_integer_divide_overflow_errors() {
-        let values = buffer![i64::MIN].into_array();
-        let result = values
-            .binary(
-                ConstantArray::new(-1i64, values.len()).into_array(),
-                Operator::Div,
-            )
-            .and_then(|a| a.execute::<PrimitiveArray>(&mut array_session().create_execution_ctx()));
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_integer_divide_errors_ignore_null_lanes() {
-        let mut ctx = array_session().create_execution_ctx();
-        let lhs = PrimitiveArray::new(buffer![10i32, 10], Validity::from_iter([false, true]))
-            .into_array();
-        let rhs = buffer![0i32, 2].into_array();
-        let result = lhs
-            .binary(rhs, Operator::Div)
-            .and_then(|a| {
-                a.execute::<RecursiveCanonical>(&mut array_session().create_execution_ctx())
-            })
-            .map(|a| a.0.into_array())
-            .unwrap();
-
-        assert_arrays_eq!(
-            result,
-            PrimitiveArray::from_option_iter([None, Some(5i32)]),
-            &mut ctx
-        );
-    }
-
-    #[test]
-    fn test_integer_errors_ignore_null_lanes() {
-        let mut ctx = array_session().create_execution_ctx();
-        let values = PrimitiveArray::new(buffer![u8::MAX, 1], Validity::from_iter([false, true]))
-            .into_array();
-        let result = values
-            .binary(
-                ConstantArray::new(1u8, values.len()).into_array(),
-                Operator::Add,
-            )
-            .and_then(|a| {
-                a.execute::<RecursiveCanonical>(&mut array_session().create_execution_ctx())
-            })
-            .map(|a| a.0.into_array())
-            .unwrap();
-
-        assert_arrays_eq!(
-            result,
-            PrimitiveArray::from_option_iter([None, Some(2u8)]),
-            &mut ctx
-        );
-    }
-
-    #[test]
-    fn test_integer_array_array_errors_on_valid_lanes() {
-        let lhs = PrimitiveArray::new(
-            buffer![u8::MAX, 1, u8::MAX],
-            Validity::from_iter([false, true, true]),
-        )
-        .into_array();
-        let rhs = buffer![1u8, 1, 1].into_array();
-        let result = lhs
-            .binary(rhs, Operator::Add)
-            .and_then(|a| a.execute::<PrimitiveArray>(&mut array_session().create_execution_ctx()));
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_present_nullable_constant_preserves_nullable_output() {
-        let mut ctx = array_session().create_execution_ctx();
-        let values = buffer![1u8, 2].into_array();
-        let result = values
-            .binary(
-                ConstantArray::new(Some(1u8), values.len()).into_array(),
-                Operator::Add,
-            )
-            .and_then(|a| a.execute::<PrimitiveArray>(&mut array_session().create_execution_ctx()))
-            .unwrap();
-
-        assert_arrays_eq!(
-            result,
-            PrimitiveArray::from_option_iter([Some(2u8), Some(3)]),
-            &mut ctx
-        );
-    }
-}
